Blog-Django/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.7
gunicorn
django-ckeditor==6.1.0
django-js-asset==1.2.2
Pillow==8.3.2
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.2
whitenoise==5.3.0
django-crispy-forms==1.11.2